refactor(desktop): move display scheduling into conversation ownership
Keep frame cleanup within the renderer architecture boundary without widening legacy budgets.

Generated-by: OpenAI Codex
