+++ apps/desktop/src/renderer/features/conversation/controller/display-frame-scheduler.ts
@@ -1,0 +1,35 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+export function createConversationDisplayFrameScheduler(): ((callback: () => void) => void) | undefined {
+  if (typeof requestAnimationFrame !== 'function') return undefined;
+  return (callback) => {
+    let pending = true;
+    const run = () => {
+      if (!pending) return;
+      pending = false;
+      // Hidden windows can suspend frames while the timeout keeps flushing.
+      cancelAnimationFrame(frameId);
+      window.clearTimeout(timeoutId);
+      callback();
+    };
+    const frameId = requestAnimationFrame(run);
+    const timeoutId = window.setTimeout(run, 100);
+  };
+}
